+++ index.js
@@ -161,6 +161,12 @@ exports.format = function(name, fmt){
 exports.format('combined', ':remote-addr - :remote-user [:date] ":method :url HTTP/:http-version" :status :res[content-length] ":referrer" ":user-agent"')
 
 /**
+ * Apache common log format.
+ */
+
+exports.format('common', ':remote-addr - :remote-user [:date] ":method :url HTTP/:http-version" :status :res[content-length]')
+
+/**
  * Default format.
  */