+++ History.md
@@ -3,6 +3,7 @@ unreleased
 
   * Add `:remote-user` token
   * Add `combined` log format
+  * Add `common` log format
   * Remove non-standard grey color from `dev` format
 
 1.1.1 / 2014-05-20